Fixed all the errors and warnings from dialyzer for now
I commented out the code to make this work.

Aside:
Dialyzer is optimistic, it is not strict like Haskell’s compiler. For
some reason, this feels weird.
In addition to this, there does not seem to be much documentation
online explaining its use. I figured out how to specify the basic types
but I still need to figure out how to specify the polymorphic types and
data structures like sets() and dicts().