Implements election timeout and configuration validation
The configuration validation code throws an error if the config file is
not valid.

The election timeout code works. If you run a single server, it becomes
a candidate and restarts the election over and over as it has no peers

Added some more notes in implementation_notes.md w.r.t the testing and
gen_statem

The call back mode for gen_statem was modified to include state_enter
events.