Candidate now initiates the election
Upon election timeout, the candidate sends out request votes request to
all the peers. It retries the request for unresponsive peers at
heartbeat intervals.

Election timeout for the current term is calculated when the election
for that term is started.

Refactored the timeout code into functions so that the time actions are
easier to read.

Extended the raft_state record with votes_received & votes_rejected to
keep track of responses to request votes