Initial commit of leader election
Manually tested it on 3 and 5 nodes by starting the app in different
terminals. Works as expected even during leader failures. Haven’t
tested the follower failures

Currently, it does not handle any client requests. So the indices &
terms in append entries request are 0.

A Candidate retries the request votes on unresponsive peers after every
heartbeat. It also steps down as a candidate if it receives a request
votes request from a candidate with higher term.

The leader simply maintains it’s authority by sending no-op heartbeats