Fix a leader election bug w.r.t counting votes
I was not counting the votes from self & the vote from the peer (from
request votes response) properly.

I was using ‘+1’ instead of ‘+2’. This showed up as a bug when I had
one node not responding to messages in a cluster of 3. The majority was
never achieved due to this bug.

The bug was found accidentally when playing with test_proxy. I had one
server started in test configuration and the others in normal
configuration. The server in test configuration was not able to reply
to any messages from other servers as it was sending messages to the
test_proxy on those nodes. They do not exist as they are not in test
configuration and hence were sent into vacuum.