Using mnesia as the storage engine for raft's replication log
Storage engine for replication log needs to provide atomicity (for
reads & writes) & durability. Mnesia is an ACID db, so it meets our
requirements. We do not care about consistency (nilva ensures
consistency amongst tables) & isolation (we have a sequential access by
design) as we do not need them here. Even distributed parts of mnesia
db are useless to us as we use it in local only config.

We do not need a separate nilva_log_server process as mnesia already
runs in a separate process.

Converted the nilva_log_server from gen_server to a normal module
called nilva_replication_log.erl. The module separates the
nilva_raft_fsm from implementation details of the storage engine.

Mnesia will be used to create tables locally on the node and will not
be replicating the tables to any other nodes as such.

Mnesia’s disc_copies tables have some limitations on table size but
this should not be a limitation for the initial implementation.