Fixes #7 #8
For #7, index has been made the primary key in the nilva_log_entry
table.

For #8, record_info was the cause of the bug. It is evaluated at
compile time, so all the tables were being created with the same
attributes of nilva_log_entry. This was not found until now as I had
been using the same schema & tables without deleting & creating them.
This was fixed along with some addition of error handling code